requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
colorgram.py==1.2.0
idna==3.11
numpy==2.3.4
pandas==2.3.3
pandas-stubs==2.3.2.250926
pillow==12.0.0
prettytable==3.16.0
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
